requirements.txt
astroid==2.5.6
click==7.1.2
Flask==1.1.2
gunicorn==20.1.0
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.23
protobuf==3.15.7
pylint==2.8.2
python-dotenv==0.17.1
six==1.15.0
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
